docs how-to-works.md
